+++ UltimateCastbars/Backend/Core/Static/BarUpdate_Helpers.lua
@@ -243,6 +243,33 @@ function BarUpdate_API:UpdateText(unit)
                 usedShadowColour = tagOptions.shadowColour
             end
 
+            if tagOptions.sizeControl.widthControl.enable then
+                local width
+                if tagOptions.sizeControl.widthControl.type == "percentage" then
+                    width = bar:GetWidth() * (tagOptions.sizeControl.widthControl.perValue / 100)
+                else
+                    width = tagOptions.sizeControl.widthControl.customValue
+                end
+                bar.texts[key]:SetWidth(width)
+            else
+                bar.texts[key]:SetWidth(9999999)
+            end
+
+            if tagOptions.sizeControl.heightControl.enable then
+                local height
+                if tagOptions.sizeControl.heightControl.type == "percentage" then
+                    height = bar:GetHeight() * (tagOptions.sizeControl.heightControl.perValue / 100)
+                else
+                    height = tagOptions.sizeControl.heightControl.customValue
+                end
+                bar.texts[key]:SetHeight(height)
+                bar.texts[key]:SetNonSpaceWrap(false)
+            else
+                bar.texts[key]:SetHeight(9999999)
+                bar.texts[key]:SetNonSpaceWrap(false)
+            end
+            bar.texts[key]:SetWordWrap(tagOptions.sizeControl.heightControl.wrapText)
+
             local fs = bar.texts[key]
             fs:SetJustifyH(tagOptions.justify)
             fs:SetPoint(tagOptions.anchorFrom, textFrame, tagOptions.anchorTo, tagOptions.textOffsetX, tagOptions.textOffsetY)