+++ UltimateCastbars/Backend/DefaultDB/Target_DB.lua
@@ -179,7 +179,23 @@ local target_db = {
 
             extraOptions = {
                 useClassColour = true,
-            }
+            },
+
+            sizeControl = {
+                widthControl = {
+                    enable = false,
+                    type = "percentage",
+                    perValue = 100,
+                    customValue = 300,
+                },
+                heightControl = {
+                    enable = false,
+                    wrapText = false,
+                    type = "percentage",
+                    perValue = 100,
+                    customValue = 300,
+                },
+            },
 
         },
         textList = {
@@ -219,7 +235,23 @@ local target_db = {
 
                 extraOptions = {
                     useClassColour = true,
-                }
+                },
+
+                sizeControl = {
+                    widthControl = {
+                        enable = false,
+                        type = "percentage",
+                        perValue = 100,
+                        customValue = 300, 
+                    },
+                    heightControl = {
+                        enable = false,
+                        wrapText = false,
+                        type = "percentage",
+                        perValue = 100,
+                        customValue = 300,
+                    },
+                },
             },
             tag2 = {
                 name = "Timer (decrease)",
@@ -256,7 +288,24 @@ local target_db = {
 
                 extraOptions = {
                     useClassColour = true,
-                }
+                },
+
+                sizeControl = {
+                    widthControl = {
+                        enable = false,
+                        type = "percentage",
+                        perValue = 100,
+                        customValue = 300, 
+                    },
+                    heightControl = {
+                        enable = false,
+                        wrapText = false,
+                        type = "percentage",
+                        perValue = 100,
+                        customValue = 300,
+                    },
+                },
+
             },
             tag3 = {
                 name = "Timer (increase)",
@@ -293,7 +342,24 @@ local target_db = {
 
                 extraOptions = {
                     useClassColour = true,
-                }
+                },
+
+                sizeControl = {
+                    widthControl = {
+                        enable = false,
+                        type = "percentage",
+                        perValue = 100,
+                        customValue = 300, 
+                    },
+                    heightControl = {
+                        enable = false,
+                        wrapText = false,
+                        type = "percentage",
+                        perValue = 100,
+                        customValue = 300,
+                    },
+                },
+
             },
             tag4 = {
                 name = "Cancelled",
@@ -330,7 +396,23 @@ local target_db = {
 
                 extraOptions = {
                     useClassColour = true,
-                }
+                },
+
+                sizeControl = {
+                    widthControl = {
+                        enable = false,
+                        type = "percentage",
+                        perValue = 100,
+                        customValue = 300, 
+                    },
+                    heightControl = {
+                        enable = false,
+                        wrapText = false,
+                        type = "percentage",
+                        perValue = 100,
+                        customValue = 300,
+                    },
+                },
             },
             tag5 = {
                 name = "Interrupted By",
@@ -367,7 +449,24 @@ local target_db = {
                 
                 extraOptions = {
                     useClassColour = true,
+                },
+
+                sizeControl = {
+                    widthControl = {
+                        enable = false,
+                        type = "percentage",
+                        perValue = 100,
+                        customValue = 300,
+                    },
+                    heightControl = {
+                        enable = false,
+                        wrapText = false,
+                        type = "percentage",
+                        perValue = 100,
+                        customValue = 300,
+                    },
                 }
+
             },
         },
     },